fix(computer-use): sign the helper in afterPack so notarization stays valid
Signing the nested helper in afterSign broke packaging two ways: changing it
after the app was signed invalidated the app's resource seal, and re-sealing
the app then broke electron-builder's staple (its notarization ticket was
prepared against the pre-afterSign signature). Move the helper signing to
afterPack — before electron-builder signs the app — so the app-level signature
seals the helper's final signature and electron-builder does notarize + staple
untouched. Verified end to end on a notarized build: app spctl accepted,
codesign --deep --strict passes, and both the app and dmg are stapled.
