+++ electron-builder.yml
@@ -31,9 +31,11 @@ asarUnpack:
   # jieba-wasm reads its .wasm via fs.readFileSync(__dirname/...); unpack it so
   # the file resolves on disk rather than from inside the asar archive.
   - node_modules/jieba-wasm/**
-# Re-sign the nested Computer Use helper (timestamped, hardened) after the app is
-# signed and before notarization — electron-builder doesn't deep-sign extraResources.
-afterSign: scripts/after-sign.cjs
+# Sign the nested Computer Use helper BEFORE electron-builder signs the app, so
+# the app-level signature seals its final signature and the notarize + staple
+# stay valid. Doing it in afterSign breaks the app's resource seal and the staple
+# ticket. electron-builder doesn't deep-sign extraResources, so this signature sticks.
+afterPack: scripts/after-pack.cjs
 mac:
   # Signed + notarized with a real Developer ID (Team 43PH2NJ5C4). Hardened
   # Runtime is mandatory for notarization, and it makes dyld enforce Library
@@ -68,8 +70,8 @@ mac:
   # The signed Computer Use helper ships inside Resources. It's macOS-only, so it
   # lives under mac.extraResources (appended to the common list) rather than the
   # top-level one — Windows/Linux builds never carry it. electron-builder won't
-  # deep-sign nested extraResources, so the afterSign hook signs it (Team
-  # 43PH2NJ5C4 + hardened runtime) before notarization.
+  # deep-sign nested extraResources, so the afterPack hook signs it (Team
+  # 43PH2NJ5C4 + hardened runtime) before the app-level sign seals it.
   extraResources:
     - from: native/computer-use-helper/build/Atrium Computer Use.app
       to: Atrium Computer Use.app
